refactor: footer module
