+++ package.json
@@ -13,11 +13,12 @@
     "stylelint-config-standard": "^20.0.0",
     "stylelint-csstree-validator": "^1.8.0",
     "stylelint-scss": "^3.17.2",
-    "webpack": "^5.1.3",
+    "webpack": "^5.2.0",
     "webpack-cli": "^4.1.0"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "webpack --mode development",
+    "build": "webpack --mode production",
     "start": "npm run start-dev"
   },
   "repository": {
