+++ src/selectors.js
@@ -1,0 +1,7 @@
+export const numberButtons = document.querySelectorAll('[data-number]');
+export const operationButtons = document.querySelectorAll('[data-operation]');
+export const equalsButton = document.querySelector('[data-equals]');
+export const deleteButton = document.querySelector('[data-delete]');
+export const allClearButton = document.querySelector('[data-all-clear]');
+export const previousOperandTextElement = document.querySelector('[data-previous-operand]');
+export const currentOperandTextElement = document.querySelector('[data-current-operand]');
